[ehn] Add DMARC record
Add a DMARC record. From Wikipedia, "a DMARC policy allows a sender's domain to indicate that their emails are protected by SPF and/or DKIM, and tells a receiver what to do if neither of those authentication methods passes". Proposed policy is to log affected messages and send daily reports to postmaster@DOMAIN.TLD.